+++ src/__mocks__/react.js
@@ -1,12 +1,12 @@
 /* global window */
-const react = require('react');
+const react = require('react')
 // Resolution for requestAnimationFrame not supported in jest error :
 // https://github.com/facebook/react/issues/9102#issuecomment-283873039
-global.window = global;
+global.window = global
 
-window.addEventListener = () => {};
+window.addEventListener = () => {}
 window.requestAnimationFrame = () => {
-  throw new Error('requestAnimationFrame is not supported in Node');
-};
+  throw new Error('requestAnimationFrame is not supported in Node')
+}
 
-module.exports = react;
+module.exports = react