+++ src/__tests__/helpers/events.js
@@ -1,27 +1,19 @@
 function createClientXYObject(x, y) {
-  return { clientX: x, clientY: y };
+  return { clientX: x, clientY: y }
 }
-
-export function createStartTouchEventObject({ x = 0, y = 0, preventDefault = () => {} }) {
+const preventDefault = () => {}
+export function createTouchEventObject({ x, y, ...rest }) {
   return {
     touches: [createClientXYObject(x, y)],
     preventDefault,
-  };
+    ...rest
+  }
 }
 
-export function createMoveTouchEventObject(props) {
-  const { x = 0, y = 0, includeTouches = true, preventDefault = () => {} } = props;
-  const moveTouchEvent = {
-    changedTouches: [createClientXYObject(x, y)],
-    preventDefault,
-  };
-  if (includeTouches) moveTouchEvent.touches = [createClientXYObject(x, y)];
-  return moveTouchEvent;
-}
-
-export function createMouseEventObject({ x = 0, y = 0, preventDefault = () => {} }) {
+export function createMouseEventObject({ x, y, ...rest }) {
   return {
     ...createClientXYObject(x, y),
     preventDefault,
-  };
+    ...rest
+  }
 }